create utils file, refactor
