+++ src/components/AddItem.jsx
@@ -4,34 +4,7 @@ import { getDoc, doc, setDoc } from 'firebase/firestore';
 import { db } from '../lib/firebase';
 import { useNavigate } from 'react-router-dom';
 import toast, { Toaster } from 'react-hot-toast';
-
-const radioButtonOptions = [
-  {
-    option: 'soon',
-    value: 7,
-    defaultChecked: true,
-  },
-  {
-    option: 'kind of soon',
-    value: 14,
-    defaultChecked: false,
-  },
-  {
-    option: 'not soon',
-    value: 30,
-    defaultChecked: false,
-  },
-];
-
-function removePunctuation(string) {
-  const punctuationlessString = string
-    .toLowerCase()
-    .replace(/[^\w\s]|_/g, '')
-    .split(' ')
-    .join('');
-
-  return punctuationlessString;
-}
+import { radioButtonOptions, removePunctuation } from '../utils';
 
 //duplicate item? sets isDuplicateFound to true, otherwise false
 //isDuplicateFound boolean gates form submission to database
