+++ src/components/ListLayout.jsx
@@ -1,26 +1,13 @@
 import React, { useEffect, useState, useRef } from 'react';
-import { updateDoc, doc, deleteDoc } from 'firebase/firestore';
-
-import { db } from '../lib/firebase';
+import { setUpdateToDb, deleteItemFromDb } from '../lib/firebase';
 import { ImCross } from 'react-icons/im';
 import { RiDeleteBin6Fill } from 'react-icons/ri';
 import { calculateEstimate } from '@the-collab-lab/shopping-list-utils';
-
-// UPDATED this function to accommodate changes to multiple values on an item object
-const setUpdateToDb = async (collection, itemId, dataToUpdate) => {
-  const itemRef = doc(db, collection, itemId);
-  await updateDoc(itemRef, dataToUpdate);
-};
-//create a function to delete item from db
-const deleteItemFromDb = async (token, itemId) => {
-  await deleteDoc(doc(db, token, itemId));
-};
+import { oneDay, currentTime, within24hours, groups } from '../utils';
 
 const ListLayout = ({ items, localToken }) => {
   const [filter, setFilter] = useState('');
 
-  const currentTime = Date.now();
-  const oneDay = 86400000; //24 hours in milliseconds
   //create a reference for an input
   const inputRef = useRef(null);
 
@@ -52,22 +39,11 @@ const ListLayout = ({ items, localToken }) => {
     }
   };
 
-  //persists checked box for 24 hours
-  function within24hours(date) {
-    let timeCheck = false;
-    const gap = currentTime - date;
-    if (gap < oneDay) {
-      timeCheck = true;
-    }
-    return timeCheck;
-  }
-
-
-  function deleteButtonPressed(itemId, itemName) {
+  const deleteButtonPressed = (itemId, itemName) => {
     if (window.confirm(`Are you sure you want to delete ${itemName}?`)) {
       deleteItemFromDb(localToken, itemId);
     }
-  }
+  };
 
   // updates isActive property of item to true if item has 2+ purchases and has been purchased within calculated estimate
   // isActive is defaulted to false when item is added
@@ -88,52 +64,12 @@ const ListLayout = ({ items, localToken }) => {
       }
     });
     //suggested dependency array via React and I agree with the suggestion if anyone has thoughts on this please let me know!
-  }, [items, localToken, currentTime]);
+  }, [items, localToken]);
 
   //filters items to only display items a user is searching by via the input bar
   const filteredItems = items.filter((item) =>
     item.id.includes(filter.toLowerCase()),
   );
-
-  const groups = [
-    {
-      label: 'Soon',
-      sublabel: "We think you'll need this in less than 7 days",
-      groupFilter: (item) => {
-        return item.previousEstimate < 7 && item.isActive === true;
-      },
-      colorClass: 'bg-rose-100',
-    },
-    {
-      label: 'Kind of soon',
-      sublabel: "We think you'll need this in less than 30 days",
-      groupFilter: (item) => {
-        return (
-          item.previousEstimate >= 7 &&
-          item.previousEstimate < 30 &&
-          item.isActive === true
-        );
-      },
-      colorClass: 'bg-yellow-100',
-    },
-    {
-      label: 'Not soon',
-      sublabel: "We think you'll need this in more than 30 days",
-      groupFilter: (item) => {
-        return item.previousEstimate >= 30 && item.isActive === true;
-      },
-      colorClass: 'bg-green-100',
-    },
-    {
-      label: 'Inactive',
-      sublabel: "This item is inactive and hasn't been purchased recently",
-      groupFilter: (item) => {
-        return item.isActive === false;
-      },
-      colorClass: 'bg-gray-200',
-    },
-  ];
-
 
   return (
     <>
@@ -155,7 +91,7 @@ const ListLayout = ({ items, localToken }) => {
           </svg>
         </span>
         <input
-          className="text-black bg-violet-100 p-2 text-md bg-gray-900 rounded-md pl-10"
+          className="text-black bg-violet-100 p-2 text-md rounded-md pl-10"
           type="text"
           id="search"
           ref={inputRef}
@@ -203,13 +139,15 @@ const ListLayout = ({ items, localToken }) => {
                           name={item.id}
                           aria-label={item.itemName}
                         />
-                         <button
-                            aria-label={`delete ${item.id} button`}
-                            className="bg-blue-500 hover:bg-blue-700 text-white ml-4 font-bold py-1 px-1 rounded"
-                            onClick={() => deleteButtonPressed(item.id, item.itemName)}
-                         >
-                           <RiDeleteBin6Fill />
-                       </button>
+                        <button
+                          aria-label={`delete ${item.id} button`}
+                          className="bg-blue-500 hover:bg-blue-700 text-white ml-4 font-bold py-1 px-1 rounded"
+                          onClick={() =>
+                            deleteButtonPressed(item.id, item.itemName)
+                          }
+                        >
+                          <RiDeleteBin6Fill />
+                        </button>
                       </div>
                       <div className="px-4">{` Time until next purchase: ${item.previousEstimate}`}</div>
                       <div className="px-4">{` Total purchases: ${item.totalPurchases}`}</div>
