+++ src/components/Navigation.jsx
@@ -1,14 +1,6 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
-
-//might want to adjust these styles or take them out and replace with tailwind classes..
-const navStyles = {
-  position: 'fixed',
-  bottom: '0',
-  width: '100%',
-  textAlign: 'center',
-  backgroundColor: '#fff',
-};
+import { navStyles } from '../utils';
 
 export default function Navigation() {
   return (
